refactor: Remove unused ProductList component and update GetProductList component
